수정 필요할수있지만 1차 완성본

+++ src/main/java/com/green/greenfirstproject/boardMaster/BoardMasterMapper.java
@@ -19,7 +19,7 @@ public interface BoardMasterMapper {
 
     int patchBoardHits(long boardSeq);
 
-    Long totalCount(String keyword);
+    Integer totalCount(String keyword);
 
 //    List<BoardGetRes> getCommunityListTitle(BoardGetReq p);
 //
